Infrastructure and migrations are complete for saving a player's nemesis in the DB. Now just need to update createPlayedGame and de-activate logic.

+++ Source/UI/Models/Players/PlayerDetailsViewModel.cs
@@ -21,7 +21,7 @@ namespace UI.Models.Players
         public bool HasNemesis { get; set; }
         public int NemesisPlayerId { get; set; }
         public string NemesisName { get; set; }
-        public int LossPercentageVersusPlayer { get; set; }
+        public float LossPercentageVersusPlayer { get; set; }
         public int NumberOfGamesLostVersusNemesis { get; set; }
         public bool UserCanEdit { get; set; }
     }
